# FlaskStudent/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Student API",
    "version": "1.0"
  },
  "paths": {
    "/students/": {
      "get": {
        "summary": "Get all students",
        "responses": {
          "200": {
            "description": "A list of students"
          }
        }
      },
      "post": {
        "summary": "Create a new student",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Student created successfully"
          }
        }
      }
    },
    "/students/{student_id}": {
      "get": {
        "summary": "Get student by ID",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Student data"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "put": {
        "summary": "Update student by ID",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Student"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student updated successfully"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "delete": {
        "summary": "Delete student by ID",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted successfully"
          },
          "404": {
            "description": "Student not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Student": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "grade": {
          "type": "string"
        }
      }
    }
  }
}